finishing touches on backend

+++ Clanak/src/main/java/com/example/demo/Controller/KorisnikController.java
@@ -6,6 +6,7 @@ import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.Errors;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -29,38 +31,77 @@ public class KorisnikController {
 	@Autowired
 	KorisnikRepository korisnikRepository;
 	
-	//@PreAuthorize("hasAuthority('ROLE_USER')")
-	//@Secured("ROLE_USER")
+	
 	@GetMapping(value = "")
-	public List<Korisnik> getAll(){
-		return korisnikRepository.findAll();
+	public List<Korisnik> getAll(@RequestHeader(value="role") String acceptHeader){
+		
+		if (acceptHeader.equals("ROLE_AUTOR")) {
+			 //ovdje pocetak
+			return korisnikRepository.findAll();
+			 //ovdje kraj
+		 }
+		else {		
+				throw new AccessDeniedException("nepravilna rola");
+			}
+		
+		
 	}
 	
 	@GetMapping(value = "/{id}")
-	public Korisnik getUserById(@PathVariable(value = "id") Long id) throws NotFoundException {
-		return korisnikRepository.findById(id).orElseThrow(() -> new NotFoundException("User with given id not found!"));
-	}
+	public Korisnik getUserById(@PathVariable(value = "id") Long id,@RequestHeader(value="role") String acceptHeader) throws NotFoundException {
+		
+		if (acceptHeader.equals("ROLE_AUTOR")) {
+			 //ovdje pocetak
+			return korisnikRepository.findById(id).orElseThrow(() -> new NotFoundException("User with given id not found!"));
+			
+			 //ovdje kraj
+		 }
+		else {		
+				throw new AccessDeniedException("nepravilna rola");
+			}
+		
+		}
 	
 	@PostMapping(value = "")
-	public Korisnik createUser(@RequestBody @Valid final Korisnik korisnik, Errors errors)throws Exception {
+	public Korisnik createUser(@RequestBody @Valid final Korisnik korisnik,@RequestHeader(value="role") String acceptHeader, Errors errors)throws Exception {
 		/*if(errors.hasErrors()) {
 			throw new Exception(errors.getAllErrors().get(0).getDefaultMessage());
 		}*/
-		return korisnikRepository.save(korisnik);
+		
+		if (acceptHeader.equals("ROLE_AUTOR")) {
+			 //ovdje pocetak
+			return korisnikRepository.save(korisnik);
+			 //ovdje kraj
+		 }
+		else {		
+				throw new AccessDeniedException("nepravilna rola");
+			}
+		
 	}
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<?> deleteUser(@PathVariable(value = "id") Long id) throws NotFoundException{
-		Korisnik korisnik = korisnikRepository
-				.findById(id)
-				.orElseThrow(() -> new NotFoundException("User with given id not found!"));
-		korisnikRepository.delete(korisnik);
-		return ResponseEntity.ok().build();
+	public ResponseEntity<?> deleteUser(@PathVariable(value = "id") Long id,@RequestHeader(value="role") String acceptHeader) throws NotFoundException{
+		if (acceptHeader.equals("ROLE_AUTOR")) {
+			 //ovdje pocetak
+			Korisnik korisnik = korisnikRepository
+					.findById(id)
+					.orElseThrow(() -> new NotFoundException("User with given id not found!"));
+			korisnikRepository.delete(korisnik);
+			return ResponseEntity.ok().build();
+			 //ovdje kraj
+		 }
+		else {		
+				throw new AccessDeniedException("nepravilna rola");
+			}
+		
+		
 	}
 	
 	@PutMapping("/{id}")
-	public Korisnik updateKorisnik(@PathVariable(value = "id") Long id, @RequestBody @Valid Korisnik korisnikUpdate, Errors errors) throws NotFoundException, Exception{
-		/*if(errors.hasErrors()) 
+	public Korisnik updateKorisnik(@PathVariable(value = "id") Long id, @RequestBody @Valid Korisnik korisnikUpdate,@RequestHeader(value="role") String acceptHeader, Errors errors) throws NotFoundException, Exception{
+		if (acceptHeader.equals("ROLE_AUTOR")) {
+			 //ovdje pocetak
+			/*if(errors.hasErrors()) 
 			throw new Exception(errors.getAllErrors().get(0).getDefaultMessage());*/
 		Korisnik korisnik = korisnikRepository
 				.findById(id)
@@ -68,6 +109,14 @@ public class KorisnikController {
 		korisnik.setUsername(korisnikUpdate.getUsername());
 		korisnikUpdate = korisnikRepository.save(korisnik);
 		return korisnikUpdate;
+			 //ovdje kraj
+		 }
+		else {		
+				throw new AccessDeniedException("nepravilna rola");
+			}
+		
+		
+		
 	}  
 	
 }
